requirements.txt
Flask==3.0.3
requests==2.32.3
python-docx==1.1.2
pypdf==5.1.0
pytest==8.3.3
